CRM changes

+++ backend/LIVIS/livis/crm/views.py
@@ -13,11 +13,23 @@ from logs.utils import add_logs_util
 from rest_framework.decorators import api_view, permission_classes
 from rest_framework.permissions import AllowAny
 
+
+def check_permission(request, perm_name):
+    role = request.user.role_name
+    mp = MongoHelper().getCollection("permissions")
+    p = [i for i in mp.find()]
+    if perm_name in p[0][role]:
+        pass
+    else:
+        raise PermissionDenied
+
+
 @api_view(['PATCH'])
 @renderer_classes((TemplateHTMLRenderer,))
 @csrf_exempt
 @permission_classes((AllowAny,))
 def update_lead(request):
+    check_permission(request, "can_update_lead")
     data = json.loads(request.body)
     message, status_code = update_lead_util(data)
     if status_code == 200:
@@ -29,6 +41,7 @@ def update_lead(request):
 @api_view(['DELETE'])
 @csrf_exempt
 def delete_lead(request, id):
+    check_permission(request, "can_delete_lead")
     message, status_code = delete_lead_util(id)
     if status_code == 200:
         return HttpResponse(json.dumps({'Message': 'Success!', 'data': message}), content_type="application/json")
@@ -40,6 +53,7 @@ def delete_lead(request, id):
 @csrf_exempt
 #@permission_classes((AllowAny,))
 def create_lead(request):
+    check_permission(request, "can_create_lead")
     data = json.loads(request.body)
     message, status_code = create_lead_util(data)
     if status_code == 200:
@@ -51,6 +65,7 @@ def create_lead(request):
 @csrf_exempt
 #@permission_classes((AllowAny,))
 def get_all_leads(request):
+    check_permission(request, "can_get_all_leads")
     leads_list = get_all_leads_util()
     return HttpResponse(json.dumps( leads_list , cls=Encoder), content_type="application/json")
 
@@ -59,6 +74,7 @@ def get_all_leads(request):
 @csrf_exempt
 #@permission_classes((AllowAny,))
 def get_single_lead(request, id):
+    check_permission(request, "can_get_single_lead")
     leads_list = get_single_lead_util(id)
     return HttpResponse(json.dumps(leads_list , cls=Encoder), content_type="application/json")
 
@@ -66,6 +82,7 @@ def get_single_lead(request, id):
 @csrf_exempt
 #@permission_classes((AllowAny,))
 def check_gst(request, gst):
+    check_permission(request, "can_check_gst")
     response = check_gst_util(gst)
     return HttpResponse(json.dumps(response, cls=Encoder), content_type="application/json")
 
@@ -74,6 +91,7 @@ def check_gst(request, gst):
 @csrf_exempt
 #@permission_classes((AllowAny,))
 def create_task(request):
+    check_permission(request, "can_create_task")
     data = json.loads(request.body)
     message, status_code = create_task_util(data)
     if status_code == 200:
@@ -85,6 +103,7 @@ def create_task(request):
 @csrf_exempt
 #@permission_classes((AllowAny,))
 def get_all_tasks(request, lead_id):
+    check_permission(request, "can_get_all_tasks")
     tasks_list = get_all_tasks_util(lead_id)
     return HttpResponse(json.dumps( tasks_list , cls=Encoder), content_type="application/json")
 
@@ -92,6 +111,7 @@ def get_all_tasks(request, lead_id):
 @csrf_exempt
 #@permission_classes((AllowAny,))
 def get_single_task(request, id):
+    check_permission(request, "can_get_single_task")
     tasks_list = get_single_task_util(id)
     return HttpResponse(json.dumps(tasks_list , cls=Encoder), content_type="application/json")
 
@@ -100,6 +120,7 @@ def get_single_task(request, id):
 @csrf_exempt
 #@permission_classes((AllowAny,))
 def update_task(request):
+    check_permission(request, "can_update_task")
     data = json.loads(request.body)
     message, status_code = update_task_util(data)
     if status_code == 200:
@@ -112,6 +133,7 @@ def update_task(request):
 @csrf_exempt
 #@permission_classes((AllowAny,))
 def delete_task(request, id):
+    check_permission(request, "can_delete_task")
     message, status_code = delete_task_util(id)
     if status_code == 200:
         return HttpResponse(json.dumps({'Message': 'Success!', 'data': message}), content_type="application/json")
@@ -123,6 +145,7 @@ def delete_task(request, id):
 @csrf_exempt
 #@permission_classes((AllowAny,))
 def create_todo(request):
+    check_permission(request, "can_create_todo")
     data = json.loads(request.body)
     message, status_code = create_todo_util(data)
     if status_code == 200:
@@ -134,6 +157,7 @@ def create_todo(request):
 @csrf_exempt
 #@permission_classes((AllowAny,))
 def get_all_todo(request, task_id):
+    check_permission(request, "can_get_all_todo")
     print(task_id)
     leads_list = get_all_todo_util(task_id)
     return HttpResponse(json.dumps(leads_list , cls=Encoder), content_type="application/json")
@@ -143,6 +167,7 @@ def get_all_todo(request, task_id):
 @csrf_exempt
 #@permission_classes((AllowAny,))
 def update_todo(request):
+    check_permission(request, "can_update_todo")
     data = json.loads(request.body)
     message, status_code = update_todo_util(data)
     if status_code == 200:
@@ -155,6 +180,7 @@ def update_todo(request):
 @csrf_exempt
 #@permission_classes((AllowAny,))
 def delete_todo(request, task_id, todo_id):
+    check_permission(request, "can_delete_todo")
     message, status_code = delete_todo_util(task_id, todo_id)
     if status_code == 200:
         return HttpResponse(json.dumps({'Message': 'Success!', 'data': message}), content_type="application/json")
@@ -166,6 +192,7 @@ def delete_todo(request, task_id, todo_id):
 @csrf_exempt
 #@permission_classes((AllowAny,))
 def update_lead_status(request):
+    check_permission(request, "can_update_lead_status")
     data = json.loads(request.body)
     message, status_code = update_lead_status_util(data)
     if status_code == 200:
@@ -178,6 +205,7 @@ def update_lead_status(request):
 @csrf_exempt
 #@permission_classes((AllowAny,))
 def create_lead_source(request):
+    check_permission(request, "can_create_lead_source")
     data = json.loads(request.body)
     message, status_code = create_lead_source_util(data)
     if status_code == 200:
@@ -189,6 +217,7 @@ def create_lead_source(request):
 @csrf_exempt
 #@permission_classes((AllowAny,))
 def get_all_lead_source(request):
+    check_permission(request, "can_get_all_lead_source")
     leads_list = get_all_lead_source_util()
     return HttpResponse(json.dumps( leads_list , cls=Encoder), content_type="application/json")
 
@@ -197,6 +226,7 @@ def get_all_lead_source(request):
 @csrf_exempt
 #@permission_classes((AllowAny,))
 def get_single_lead_source(request, id):
+    check_permission(request, "can_get_single_lead_source")
     leads_list = get_single_lead_source_util(id)
     return HttpResponse(json.dumps(leads_list , cls=Encoder), content_type="application/json")
 
@@ -205,6 +235,7 @@ def get_single_lead_source(request, id):
 @csrf_exempt
 #@permission_classes((AllowAny,))
 def delete_lead_source(request, id):
+    check_permission(request, "can_delete_lead_source")
     message, status_code = delete_lead_source_util(id)
     if status_code == 200:
         return HttpResponse(json.dumps({'Message': 'Success!', 'data': message}), content_type="application/json")
